Make no_handle_scope_depth_ an int32_t field as that is what the accessor returns and there is check to ensure it;s value is < INT_MAX.
R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1530543002 .

+++ runtime/vm/thread.h
@@ -414,7 +414,7 @@ LEAF_RUNTIME_ENTRY_LIST(DEFINE_OFFSET_METHOD)
   int32_t no_callback_scope_depth_;
 #if defined(DEBUG)
   HandleScope* top_handle_scope_;
-  intptr_t no_handle_scope_depth_;
+  int32_t no_handle_scope_depth_;
   int32_t no_safepoint_scope_depth_;
 #endif
   VMHandles reusable_handles_;
